ANFEPA-2559: Rbel-Logger: Fixed the handling of the request-counter for EPA3-messages

+++ tiger-rbel/src/main/java/de/gematik/rbellogger/converter/RbelVauEpa3Converter.java
@@ -124,7 +124,7 @@ public class RbelVauEpa3Converter implements RbelConverterPlugin {
                       RbelElement.wrap(
                           reqCounterBytes,
                           headerElement,
-                          ByteBuffer.wrap(reqCounterBytes).getInt()))
+                          ByteBuffer.wrap(reqCounterBytes).getLong()))
                   .with(
                       "keyId",
                       RbelElement.wrap(
